Made some changes xD
