+++ src/auth/oauth/google.rs
@@ -1,7 +1,7 @@
 use crate::auth::oauth::error::OauthError;
 use crate::auth::oauth::provider::google::GoogleProvider;
 use crate::auth::oauth::scopes::google::{GoogleScope, GoogleScopes};
-use crate::auth::oauth::OAuthCallbackQuery;
+
 use crate::auth::{Role, UserInfo};
 use crate::error::ServerResponseError;
 use crate::models::datetime::Datetime;
@@ -58,82 +58,4 @@ define_oauth_client! {
             Ok::<UserInfo, OauthError>(UserInfo::from(google_user_info))
         },
     }
-}
-
-generate_endpoint! {
-    fn google_login;
-    method: get;
-    path: "/login";
-    docs: {
-        context_path: "/google",
-        tag: "oauth",
-        responses: {
-            (status = 302, description = "Redirect to Google login page"),
-        }
-    }
-    params: {
-        state: web::Data<AppState>,
-    };
-    {
-        info!("Redirecting to Google login page");
-        let oauth = state.oauth.clone();
-
-        let (auth_url, _csrf_token) = oauth.google.get_auth_url();
-
-        Ok(HttpResponse::Found()
-            .append_header(("Location", auth_url))
-            .finish())
-    }
-}
-
-generate_endpoint! {
-    fn google_callback;
-    method: get;
-    path: "/callback";
-    docs: {
-        context_path: "/google",
-        tag: "oauth",
-        responses: {
-            (status = 302, description = "Redirect to frontend"),
-        }
-    }
-    params: {
-        state: web::Data<AppState>,
-        query: web::Query<OAuthCallbackQuery>,
-        req: HttpRequest,
-    };
-    {
-        info!("Google callback received");
-
-        let oauth = state.oauth.clone();
-
-        let frontend_url = req.url_for_static("frontend").unwrap().to_string();
-
-        match oauth
-            .google
-            .exchange_code(query.code.clone(), &state.db)
-            .await
-        {
-            Ok(session) => {
-                let redirect_url = format!("{}redirect?token={}", frontend_url, session.access_token);
-                Identity::login(&req.extensions(), session.access_token.clone()).unwrap();
-
-                Ok(HttpResponse::Found()
-                    .append_header(("Location", redirect_url))
-                    .cookie(Cookie::new("token", session.access_token))
-                    .finish())
-            }
-
-            Err(err) => {
-                error!("Error exchanging code: {}", err);
-                Err(ServerResponseError::InternalError(err.to_string()))
-            }
-        }
-    }
-}
-
-pub fn google_oauth_service() -> Scope {
-    web::scope("/google")
-        .service(google_login)
-        .service(google_callback)
 }
