+++ src/auth/oauth/mod.rs
@@ -9,14 +9,14 @@ pub mod register;
 pub(crate) mod scopes;
 pub mod update;
 
-use crate::auth::oauth::github::{github_oauth_service, GithubOauth};
-use crate::auth::oauth::google::{google_oauth_service, GoogleOauth};
+use crate::auth::oauth::github::GithubOauth;
+use crate::auth::oauth::google::GoogleOauth;
 use crate::auth::oauth::local::token;
+use crate::dto::UserRegistrationRequest;
 use actix_web::guard::Acceptable;
 use actix_web::{web, Scope};
 use anyhow::Result;
-use logout::logout as logout_endpoint;
-use crate::dto::UserRegistrationRequest;
+
 use serde::Deserialize;
 use utoipa::{IntoParams, OpenApi};
 
@@ -34,52 +34,3 @@ impl Oauth {
         })
     }
 }
-
-pub fn oauth_service() -> Scope {
-    web::scope("/oauth")
-        .service(google_oauth_service())
-        .service(github_oauth_service())
-        .service(logout_endpoint)
-        .guard(Acceptable::new(mime::APPLICATION_JSON).match_star_star())
-        .service(token)
-}
-
-#[allow(dead_code)]
-#[derive(Deserialize, Debug, IntoParams)]
-pub struct OAuthCallbackQuery {
-    #[param(example = "code123")]
-    code: String,
-    #[param(example = "state123")]
-    state: String,
-}
-
-use github::__path_login as __path_github_login;
-use google::*;
-
-use crate::models::{access_token::AccessToken, refresh_token::RefreshToken};
-use local::{TokenRequest, TokenResponse, TokenResponseExample, TokenType, __path_token};
-use logout::__path_logout;
-use crate::dto::user_update_request::UserUpdateRequest;
-
-#[derive(OpenApi)]
-#[openapi(
-    paths(
-        google_login,
-        github_login,
-        token,
-        logout,
-    ),
-    components(
-        schemas(
-            AccessToken,
-            RefreshToken,
-            TokenRequest,
-            TokenResponse,
-            TokenType,
-            UserRegistrationRequest,
-            UserUpdateRequest
-        ),
-        responses(TokenResponseExample)
-    )
-)]
-pub struct OauthApi;
