+++ src/auth/users/mod.rs
@@ -1,10 +1,9 @@
+/*
 pub(crate) mod auth;
-pub(crate) mod create;
 pub(crate) mod get;
 
-use self::get::get_user_by;
 use super::{Role, UserInfo, UserInfoExampleResponses};
-use crate::auth::users::get::{GetUserBy, __path_get_user_by};
+use crate::services::user::get::{GetUserBy, __path_get_user_by};
 use actix_web::guard::Acceptable;
 use actix_web::web;
 use utoipa::OpenApi;
@@ -24,3 +23,4 @@ pub fn user_service() -> impl actix_web::dev::HttpServiceFactory {
         .guard(Acceptable::new(mime::APPLICATION_JSON).match_star_star())
         .service(get_user_by)
 }
+*/
