+++ src/models/mod.rs
@@ -6,7 +6,9 @@ use surrealdb::sql::Thing;
 pub mod access_token;
 pub mod datetime;
 pub mod refresh_token;
+pub mod session;
 pub mod thing;
+pub mod user_info;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Record {
