+++ src/models/user_info.rs
@@ -1,18 +1,13 @@
 use crate::auth::oauth::provider::{OauthProvider, OauthProviderName};
+use crate::dto::PaginationResponse;
+use crate::models::datetime::Datetime;
+use crate::models::thing::Thing;
+use crate::models::Record;
 use crate::server::db::INTERNAL_DB;
 use anyhow::bail;
 use serde::{Deserialize, Serialize};
-//use surrealdb::sql::{Datetime, Thing};
-use crate::models::datetime::Datetime;
-use crate::models::thing::Thing;
 use std::string::String;
-//use surrealdb::sql::Thing;
-use crate::dto::PaginationResponse;
-use crate::models::Record;
-use crate::services::user;
 use utoipa::{ToResponse, ToSchema};
-
-pub mod oauth;
 
 #[derive(Debug, Serialize, Deserialize, PartialOrd, Eq, PartialEq, Clone, Default, ToSchema)]
 pub enum Role {
