+++ src/services/user/get.rs
@@ -1,14 +1,9 @@
-/*
-pub(crate) mod utils;
-
-use crate::auth::users::get::utils::get_user_by_token;
-use crate::models::user_info::UserInfo;
 use crate::dto::UserInfoDTO;
-use crate::extractors::Authenticated;
+use crate::error::ServerResponseError;
+use crate::models::user_info::UserInfo;
+use crate::services::user::utils::get_user_by_token;
 use crate::AppState;
-use actix_web::web;
 use anyhow::Result;
-use helper_macros::generate_endpoint;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
 use surrealdb::Surreal;
@@ -125,39 +120,3 @@ impl IntoParams for GetUserBy {
         params
     }
 }
-
-use crate::models::user_info::UserInfoExampleResponses;
-use crate::error::ServerResponseError;
-
-generate_endpoint! {
-    fn get_user_by;
-    method: get;
-    path: "";
-    docs: {
-        params: (GetUserBy),
-        tag: "user",
-        responses: {
-            (status = 200, response = UserInfoExampleResponses),
-            (status = 401, description = "Invalid credentials"),
-            (status = 404, description = "User not found"),
-        },
-        security: [
-            ("bearer_token" = []),
-            ("cookie_session" = []),
-        ]
-    }
-    params: {
-        _auth: Authenticated,
-        state: web::Data<AppState>,
-        data: web::Query<GetUserBy>,
-    };
-    {
-        info!("Retrieving user");
-        let data = data.into_inner();
-        let db = &state.db;
-        let user = get_user_by_internal(db, &data).await?;
-
-        Ok(web::Json(user))
-    }
-}
-*/
