+++ src/services/user/mod.rs
@@ -1,18 +1,16 @@
+pub mod create;
 pub mod get;
 pub mod update;
+pub mod utils;
 
 use crate::auth::Role;
-use crate::endpoints::user::get::*;
-use crate::endpoints::user::update::*;
-use crate::extractors::Authenticated;
 use crate::models::user_info::UserInfo;
 use crate::models::user_info::UserInfoExampleResponses;
-use crate::services::user::get::{get_user_by_internal, GetUserBy};
-use crate::state::AppState;
+use crate::endpoints::user::get::*;
+use crate::endpoints::user::update::*;
+use crate::services::user::get::GetUserBy;
 use actix_web::guard::Acceptable;
 use actix_web::web;
-use helper_macros::generate_endpoint;
-use tracing::info;
 use utoipa::OpenApi;
 
 #[derive(OpenApi)]
